==> 04_analysis_55_64.Rmd <==
---
title: '03_analysis_wle'
---

```{r setup}
library(tidyverse)
library(rio)
library(demography)
library(ggflags)
library(gridExtra)
library(ggh4x)

options("scipen" = 100, "digits" = 4)

rm(list = ls())

load("outputs/data_final_55_64.RData")

```

# WLE adjusted for mortality
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Employed", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    total = n(),
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    survival = mean(survival, na.rm = T),
    cohort = max(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Contr = ifelse(Total > 0, Employed / Total * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wle_unadjusted <- table_wle
wle_unadjusted$var <- "WLE adjusted for mortality"
wle_unadjusted_male <- wle_unadjusted %>% filter(projected == 0) %>% rename("wle_unadjusted" = "WLE") %>% select("country", "cohort_two", "gender", "wle_unadjusted") %>% filter(gender == "Male population")
wle_unadjusted_female <- wle_unadjusted %>% filter(projected == 0) %>% rename("wle_unadjusted" = "WLE") %>% select("country", "cohort_two", "gender", "wle_unadjusted") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("WLE in years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wle_unadjusted_55_64.png", width = 9, height = 6)

```

# WLE adjusted for mortality and working hours
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Employed", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    Time = weighted.mean(working_hours_usual, weight = weight, na.rm = T),
    TimeN = sum(outcome, na.rm = T ),
    TimeSd = sd(working_hours_usual, na.rm = T),
    survival = mean(survival, na.rm = T),
    cohort = min(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Time = ifelse(is.na(Time), 0, Time),
         TimeSd = ifelse(is.na(TimeSd), 0, TimeSd),
         Time = ifelse(Time > 40, 1, Time / 40),
         TimeSd = TimeSd / 40)

table_wle <- table_wle %>% 
  mutate(Contr= ifelse(Total > 0, Employed / Total * Time * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar,
         Sd2   = ifelse(Time > 0, TimeSd / sqrt(TimeN), 0),
         Sdt   = Sd1 * Sd2 + Sd1 * Time^2 + Sd2 * Contr)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wle_adjusted <- table_wle
wle_adjusted$var <- "WLE adjusted for mortality and working hours"
wle_adjusted_male <- wle_adjusted %>% filter(projected == 0) %>% rename("wle_adjusted" = "WLE") %>% select("country", "cohort_two", "gender", "wle_adjusted") %>% filter(gender == "Male population")
wle_adjusted_female <- wle_adjusted %>% filter(projected == 0) %>% rename("wle_adjusted" = "WLE") %>% select("country", "cohort_two", "gender", "wle_adjusted") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("WLE in FTE years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wle_adjusted_55_64.png", width = 9, height = 6)

```

# WYL to retirement adjusted for mortality
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Retired", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    total = n(),
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    survival = mean(survival, na.rm = T),
    cohort = max(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Contr = ifelse(Total > 0, Employed / Total * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wyl_retirement <- table_wle
wyl_retirement$var <- "WYL to retirement"
wyl_retirement_male <- wyl_retirement %>% filter(projected == 0) %>% rename("wyl_retirement" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_retirement") %>% filter(gender == "Male population")
wyl_retirement_female <- wyl_retirement %>% filter(projected == 0) %>% rename("wyl_retirement" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_retirement") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("YWL in years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wyl_retirement_55_64.png", width = 9, height = 6)

```

# WYL to unemployment adjusted for mortality
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Unemployed", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    total = n(),
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    survival = mean(survival, na.rm = T),
    cohort = max(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Contr = ifelse(Total > 0, Employed / Total * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wyl_unemployment <- table_wle
wyl_unemployment$var <- "WYL to unemployment"
wyl_unemployment_male <- wyl_unemployment %>% filter(projected == 0) %>% rename("wyl_unemployment" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_unemployment") %>% filter(gender == "Male population")
wyl_unemployment_female <- wyl_unemployment %>% filter(projected == 0) %>% rename("wyl_unemployment" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_unemployment") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("YWL in years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wyl_unemployment_55_64.png", width = 9, height = 6)

```

# WYL to disability adjusted for mortality
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Ill health", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    total = n(),
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    survival = mean(survival, na.rm = T),
    cohort = max(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Contr = ifelse(Total > 0, Employed / Total * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wyl_disability <- table_wle
wyl_disability$var <- "WYL to disablity"
wyl_disability_male <- wyl_disability %>% filter(projected == 0) %>% rename("wyl_disability" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_disability") %>% filter(gender == "Male population")
wyl_disability_female <- wyl_disability %>% filter(projected == 0) %>% rename("wyl_disability" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_disability") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("YWL in years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wyl_disability_55_64.png", width = 9, height = 6)

```

# WYL to inactivity adjusted for mortality
```{r, warning = F}
table_wle <- data %>% mutate(outcome = ifelse(status == "Other", 1, 0))

table_wle <- table_wle %>% filter(!is.na(outcome)) %>%
  group_by(country, country_code, country_group, gender, cohort_two, age) %>%
  reframe(
    total = n(),
    Total = sum(1 * weight, na.rm = T),
    Employed = sum(outcome * weight, na.rm = T),
    Contr = weighted.mean(outcome, weight = weight, na.rm = T),
    TotalVar = Total * Contr * (1 - Contr),
    survival = mean(survival, na.rm = T),
    cohort = max(cohort)) %>%
  unique()

table_wle <- table_wle %>% 
  mutate(Contr = ifelse(Total > 0, Employed / Total * survival, 0),
         Sd1   = (Contr * (1 - Contr)) / TotalVar)

table_wle <- table_wle %>% 
  group_by(country, gender, cohort_two) %>%
  mutate(projected = 0) %>%
  complete(age = 55:64) %>% 
  mutate(projected = ifelse(is.na(projected), 1, 0)) %>%
  arrange(country, gender, age, cohort_two) %>% 
  filter(!is.na(cohort_two))

table_wle <- table_wle %>% group_by(country, gender) %>% fill(names(table_wle), .direction = c("down"))

table_wle <- table_wle %>% 
  select(country, country_code, country_group, gender, cohort, cohort_two, Contr, Sd1, projected) %>%
  group_by(country, country_code, country_group, gender, cohort_two) %>% 
  summarise(WLE = sum(Contr), Var = sum(Sd1), cohort = cohort, projected = projected) %>%
  filter(cohort == min(cohort)) %>%
  mutate(low = WLE - 1.96 * sqrt(Var), hig = WLE + 1.96 * sqrt(Var)) %>%
  unique()

wyl_inactivity <- table_wle
wyl_inactivity$var <- "WYL to inactivity"
wyl_inactivity_male <- wyl_inactivity %>% filter(projected == 0) %>% rename("wyl_inactivity" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_inactivity") %>% filter(gender == "Male population")
wyl_inactivity_female <- wyl_inactivity %>% filter(projected == 0) %>% rename("wyl_inactivity" = "WLE") %>% select("country", "cohort_two", "gender", "wyl_inactivity") %>% filter(gender == "Female population")

```

```{r, warning = F}
ggplot(table_wle, aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code)) +
  facet_grid(gender ~ country_group) +
  geom_point(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_ribbon(aes(ymin = low, ymax = hig, alpha = 1)) +
  geom_line(data = table_wle %>% filter(projected == 0), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code),
            size = 0.3) +
  geom_line(data = table_wle, 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 0.4, linetype = "dotted") +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort == min(cohort)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  geom_flag(data = table_wle %>% group_by(country_code, gender) %>% filter(cohort_two == max(cohort_two)), 
            aes(x = as.factor(cohort_two), y = WLE, group = country_code, country = country_code), 
            size = 3) +
  theme_linedraw() +
  theme(panel.grid.major.x = element_blank(),
        panel.grid.minor.x = element_blank(),
        panel.grid.minor.y = element_blank(),
        axis.line = element_line(),
        axis.ticks = element_line(),
        legend.position = "none",
        axis.text.x = element_text(angle = 45, vjust = 1, hjust=1, size=7)) +
  guides(x = "axis_minor", y = "axis_minor") +
  coord_cartesian(ylim = c(0, 8)) +
  scale_y_continuous("YWL in years (from age 55 to 64)", breaks = seq(-10, 10, 1)) +
  scale_x_discrete("Birth cohort", labels = c("", "'29/'30", "", "'33/'34", "", "'37/'38", "", "'41/'42", "", "'45/'46", "", "'49/'50", "", "'53/'54", "", "'57/'58", "", "'61/'62", "", "'65/'66")) +
  scale_alpha(range = c(0.01, 0.1))
ggsave("outputs/wyl_inactivity_55_64.png", width = 9, height = 6)

```


```{r}
library(Hmisc)

correlations_male <- cbind(wle_unadjusted_male$wle_unadjusted, wle_adjusted_male$wle_adjusted, wyl_retirement_male$wyl_retirement, wyl_unemployment_male$wyl_unemployment, wyl_disability_male$wyl_disability, wyl_inactivity_male$wyl_inactivity)

rcorr(as.matrix(correlations_male))

correlations_female <- cbind(wle_unadjusted_female$wle_unadjusted, wle_adjusted_female$wle_adjusted, wyl_retirement_female$wyl_retirement, wyl_unemployment_female$wyl_unemployment, wyl_disability_female$wyl_disability, wyl_inactivity_female$wyl_inactivity)

rcorr(as.matrix(correlations_female))

```

```{r}
```
